Changed the list to ZSet, and used the timestamp as the score. The last joined member will scan the score of the member at the head of the queue, and if it expires, the head of the team will be popped up.
Compared with the list, ZSet implements order, and can set the expiration time for each member, to prevent the queue from being unable to expire or the head of the queue always exist causing abnormalities.

For example, after thread B executes the subscription, thread B dies and does not perform the subsequent pop-up operation. However, there are a steady stream of new threads waiting to be queued, the expiration time of the list is always refreshed, and the uuid of thread B will always exist. At the head of the team.